Restructure and rename directories and files
Restructured and renamed project directories and files as follows;

-Made two directories, "client" and "server"  in the root directory

-Moved frontend related contents within the root directory to client
 directory

-Moved backend related contents within the root directory to server
 directory

-Made the following directories under client directory; css, images, js
 and pages

-Moved files within client directory to their appropriate client
 sub-directories

-Made the following directories under server directory; controllers,
 helpers, routes, models and tests

-Moved files within server directory to their appropriate server
 sub-directories

-Renamed directories and files following Airbnb style guide
-Updated links/paths where files, modules or classes are called or
 imported for consumption

[storyID: #159287875]
