+++ .babelrc
@@ -1,4 +1,4 @@
 
 {
-    "presets": ["env"]
+    "presets": ["env", "es2015"]
 }
