+++ client/js/diary-functions.js
