+++ client/pages/dashboard.html
@@ -4,11 +4,11 @@
    <head>
         <meta charset = "UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1">
-        <link rel="shortcut icon" href="images/logo.png" type="image/x-icon">
+        <link rel="shortcut icon" href="../images/logo.png" type="image/x-icon">
         <meta name="description" content="My Diary Application">
         <title>MyDiary</title>
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-        <link rel = "stylesheet" type = "text/css" href = "css/style.css">
+        <link rel = "stylesheet" type = "text/css" href = "../css/style.css">
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
    </head>
    
@@ -16,11 +16,11 @@
       <main>
           <header>
             <div class="nav" id="nav">
-                <a href="#toEntries"><img src="images/logo.png" id="app-logo"/>
+                <a href="#toEntries"><img src="../images/logo.png" id="app-logo"/>
                 <span id="app-name">MyDiary</span></a>          
                 <a href="#toEntries" id="active">Dashboard</a>
                 <a href="#toProfile">User Profile</a>
-                <a href="../index.html">Sign out</a>
+                <a href="index.html">Sign out</a>
                 <a href="javascript:void(0);" class="icon" onclick="toggleMenu(this)">
                     <!--i class="fa fa-bars"></i-->
                     <div class="bar1"></div>
@@ -191,7 +191,7 @@
                     <center>Registered Email <strong>erozonachi.james@gmail.com</strong></center>
                     <center>Your have <strong>100 entries</strong> since registration</center>
                   </div>
-                  <img src="images/avatar.jpg" alt="Eneh James" class="avatar center">
+                  <img src="../images/avatar.jpg" alt="Eneh James" class="avatar center">
                   <div>
                     <form id="picForm" enctype="multipart/form-data" onsubmit="return false;">
                       <center><div class="space-up">
@@ -210,9 +210,9 @@
                         </label></div>
                 </div>
       </main>
-      <script type="text/javascript" src="js/interactive.js"></script>
-      <script type="text/javascript" src="js/jquery.js"></script>
-      <script type="text/javascript" src="js/diary-functions.js"></script>
+      <script type="text/javascript" src="../js/interactive.js"></script>
+      <script type="text/javascript" src="../js/jquery.js"></script>
+      <script type="text/javascript" src="../js/diary-functions.js"></script>
    </body>
    
 </html>
