+++ client/pages/index.html
@@ -4,11 +4,11 @@
    <head>
         <meta charset = "UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1">
-        <link rel="shortcut icon" href="user/images/logo.png" type="image/x-icon">
+        <link rel="shortcut icon" href="../images/logo.png" type="image/x-icon">
         <meta name="description" content="MyDiary Application">
         <title>MyDiary</title>
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-        <link rel = "stylesheet" type = "text/css" href = "user/css/style.css">
+        <link rel = "stylesheet" type = "text/css" href = "../css/style.css">
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
    </head>
    
@@ -16,12 +16,12 @@
       <main>
           <header>
             <div class="nav" id="nav">
-                <a href="javascript:void(0);"><img src="user/images/logo.png" id="app-logo"/>
+                <a href="javascript:void(0);"><img src="../images/logo.png" id="app-logo"/>
                 <span id="app-name">MyDiary</span></a>          
                 <a href="javascript:void(0);" id="active">Home</a>
                 <a href="signup.html">Sign Up</a>
                 <a href="signin.html">Sign In</a>
-                <a href="user/dashboard.html">Dashboard</a>
+                <a href="dashboard.html">Dashboard</a>
                 <a href="javascript:void(0);" class="icon" onclick="toggleMenu(this)">
                     <!--i class="fa fa-bars"></i-->
                     <div class="bar1"></div>
@@ -85,11 +85,11 @@
                 </div>
             </form>
           </section>
-          <section><article><img id="andela-logo" src="user/images/andela.png"/><span class="heavy">Developer Challenge</span></article></section>
+          <section><article><img id="andela-logo" src="../images/andela.png"/><span class="heavy">Developer Challenge</span></article></section>
           <section><article id="footer">All rights reserved, <strong>MyDiary</strong> &copy; 2018</article></section>
       </main>
-      <script type="text/javascript" src="user/js/interactive.js"></script>
-      <script type="text/javascript" src="user/js/jquery.js"></script>
+      <script type="text/javascript" src="../js/interactive.js"></script>
+      <script type="text/javascript" src="../js/jquery.js"></script>
    </body>
    
 </html>
