+++ client/pages/signup.html
@@ -4,11 +4,11 @@
    <head>
         <meta charset = "UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1">
-        <link rel="shortcut icon" href="user/images/logo.png" type="image/x-icon">
+        <link rel="shortcut icon" href="../images/logo.png" type="image/x-icon">
         <meta name="description" content="MyDiary Application">
         <title>MyDiary</title>
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-        <link rel = "stylesheet" type = "text/css" href = "user/css/style.css">
+        <link rel = "stylesheet" type = "text/css" href = "../css/style.css">
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
    </head>
    
@@ -16,7 +16,7 @@
       <main>
           <header>
                 <div class="nav" id="nav">
-                    <a href="index.html"><img src="user/images/logo.png" id="app-logo"/>
+                    <a href="index.html"><img src="../images/logo.png" id="app-logo"/>
                     <span id="app-name">MyDiary</span></a>          
                     <a href="index.html">Home</a>
                     <a href="#" id="active">Sign Up</a>
@@ -68,8 +68,8 @@
                 </form>
           </section>
       </main>
-      <script type="text/javascript" src="user/js/interactive.js"></script>
-      <script type="text/javascript" src="user/js/jquery.js"></script>
+      <script type="text/javascript" src="../js/interactive.js"></script>
+      <script type="text/javascript" src="../js/jquery.js"></script>
    </body>
    
 </html>
