+++ server/helpers/outputlib/formatResponse.js
